tests: added tests for scim client

+++ canaille/backends/ldap/backend.py
@@ -396,13 +396,16 @@ class LDAPBackend(Backend):
 
     def delete(self, instance):
         # run the instance delete callback if existing
-        if hasattr(instance, "delete"):
-            instance.delete()
+        delete_callback = instance.delete() if hasattr(instance, "delete") else iter([])
+        next(delete_callback, None)
 
         try:
             self.connection.delete_s(instance.dn)
         except ldap.NO_SUCH_OBJECT:
             pass
+
+        # run the instance delete callback again if existing
+        next(delete_callback, None)
 
     def reload(self, instance):
         # run the instance reload callback if existing
